requirements.txt
mysql-connector-python==8.0.23
redis==5.2.1
flask==3.1.0
flask_limiter==3.10.1
gunicorn==23.0.0
passlib==1.7.4
werkzeug==3.1.3
cryptography==44.0.0
passlib==1.7.4